+++ src/frontend/components/FieldsSectionCard/utils.js
@@ -1,18 +1,31 @@
 import React from 'react'
+
 import Select from 'react-select'
 import FieldLabel from '../FieldLabel'
+import Input from '../Input'
 
 const INPUT_MAP = {
-  'Select': Select,
+  Select,
+  TextInput: Input,
+  DateInput: Input,
+  EmailInput: Input,
+  NumberInput: Input,
+  RangeInput: Input,
+  TimeInput: Input,
+  CheckboxInput: Input,
 }
 
-const generateCardInput = field => {
+const generateCardInput = ({ field, fieldLabelStyle, fieldStyle }) => {
   const { key, label, inputComponent, inputProps } = field
+
   const InputComponent = INPUT_MAP[inputComponent]
+  if (!InputComponent) return null
 
   return (
-    <div key={`field-section-card-${ key }-input`}>
-      <FieldLabel label={label}/>
+    <div key={`field-section-card-${key}-${label}-input`} style={fieldStyle}>
+      <FieldLabel id={`field-${key}-label`} labelStyle={fieldLabelStyle}>
+        {label}
+      </FieldLabel>
       <InputComponent {...inputProps} />
     </div>
   )