+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - Started: 3-11-20
 
+### Added
+- `Orion`: Created new Pathways Accounts page that features BOM modal with the `useBom` hook
+- `Polaris`: Built reusable components - `Sidebar`, `Modal`, `FieldsSectionCard`, `FieldLabel` and `Input`
+
 ### Changed
 - `Polaris`: upgraded all packages to the latest minor/patch versions (no major upgrades needed)
 - `Orion`: Refactored the `useAquila` hook to fetch filter options and placard configs separately.